refactor(discord): extract media helpers (not the status data) to cut file size
The previous fix moved ROTATING_STATUSES into a new constants.ts to get the adapter
under the 1200-line cap, but that flagged SonarCloud's duplication gate: the 100-entry
presence table reads as duplicated once literals are normalized, and moving an array
that already existed on develop turned it into 'new code'. jscpd (token-exact) finds
zero real clones — it's a data table, not logic duplication.

Instead, keep the status array inline (unchanged → not new code) and extract the
genuinely-cohesive inbound-media helpers (extractDiscordMedia, downloadDiscordAttachment)
into media.ts, matching the other bots' structure. adapter.ts is now 1175 lines; the
unused mediaKindFromMime import is dropped. All 391 bot tests pass.

Entire-Checkpoint: 0afe52e3abf7
