+++ apps/bots/discord/src/adapter.ts
@@ -28,7 +28,6 @@ import {
   handleStreamingChat,
   type IncomingMedia,
   type MediaOutcome,
-  mediaKindFromMime,
   type OutboundAttachment,
   type PlatformName,
   type RichMessage,
@@ -39,6 +38,7 @@ import {
 } from "@gaia/shared";
 import {
   ActionRowBuilder,
+  ActivityType,
   ButtonBuilder,
   ButtonStyle,
   type ChatInputCommandInteraction,
@@ -51,60 +51,123 @@ import {
   MessageFlags,
   Partials,
 } from "discord.js";
-import {
-  ROTATING_STATUSES,
-  STATUS_ROTATION_INTERVAL_MS,
-  TYPING_REFRESH_INTERVAL_MS,
-} from "./constants";
+import { downloadDiscordAttachment, extractDiscordMedia } from "./media";
 
-/** A Discord attachment/sticker normalised for the shared media pipeline. */
-interface DiscordExtractedMedia {
-  /** Public CDN URL for the bytes; empty for stickers (never downloaded). */
-  url: string;
-  media: Omit<IncomingMedia, "caption">;
-}
+// ---------------------------------------------------------------------------
+// Rotating presence statuses
+// ---------------------------------------------------------------------------
 
-/**
- * Maps a Discord message's first attachment (or sticker) onto the shared
- * {@link IncomingMedia} shape, or null when the message carries no media.
- * Optional chaining keeps it safe against partial test mocks that omit the
- * `attachments` / `stickers` collections.
- */
-function extractDiscordMedia(message: Message): DiscordExtractedMedia | null {
-  const attachment = message.attachments?.first?.();
-  if (attachment) {
-    const mimeType = attachment.contentType ?? "application/octet-stream";
-    const kind = mediaKindFromMime(mimeType);
-    return {
-      url: attachment.url,
-      media: {
-        kind,
-        isVoiceNote: message.flags?.has(MessageFlags.IsVoiceMessage) ?? false,
-        mimeType,
-        filename: kind === "document" ? attachment.name : undefined,
-      },
-    };
-  }
-  if ((message.stickers?.size ?? 0) > 0) {
-    return {
-      url: "",
-      media: { kind: "sticker", isVoiceNote: false, mimeType: "image/png" },
-    };
-  }
-  return null;
-}
+const ROTATING_STATUSES: { type: ActivityType; name: string }[] = [
+  { type: ActivityType.Watching, name: "over your goals" },
+  { type: ActivityType.Listening, name: "to your inner procrastinator" },
+  { type: ActivityType.Playing, name: "personal assistant to legends" },
+  { type: ActivityType.Competing, name: "in the productivity Olympics" },
+  { type: ActivityType.Watching, name: "your productivity soar" },
+  { type: ActivityType.Listening, name: "to the sound of getting things done" },
+  { type: ActivityType.Playing, name: "life admin simulator" },
+  { type: ActivityType.Competing, name: "against your past self" },
+  { type: ActivityType.Watching, name: "for tasks worth automating" },
+  { type: ActivityType.Listening, name: "to great ideas happen" },
+  { type: ActivityType.Playing, name: "the long game with you" },
+  { type: ActivityType.Competing, name: "for the title of most helpful AI" },
+  { type: ActivityType.Watching, name: "your potential unfold" },
+  { type: ActivityType.Listening, name: "to your todo list grow" },
+  { type: ActivityType.Playing, name: "chess with your calendar" },
+  { type: ActivityType.Competing, name: "in the task completion marathon" },
+  { type: ActivityType.Watching, name: "the chaos become clarity" },
+  { type: ActivityType.Listening, name: "to your ambitions" },
+  { type: ActivityType.Playing, name: "catch-up with your goals" },
+  { type: ActivityType.Watching, name: "your habits build momentum" },
+  { type: ActivityType.Listening, name: "to your 3am ideas" },
+  { type: ActivityType.Playing, name: "scheduler extraordinaire" },
+  { type: ActivityType.Competing, name: "with yesterday's you" },
+  { type: ActivityType.Watching, name: "your dreams take shape" },
+  { type: ActivityType.Listening, name: "to the rhythm of your workflow" },
+  { type: ActivityType.Playing, name: "productivity coach unlocked" },
+  { type: ActivityType.Watching, name: "out for your deadlines" },
+  { type: ActivityType.Listening, name: "to your best ideas yet" },
+  { type: ActivityType.Playing, name: "your favorite AI companion" },
+  { type: ActivityType.Competing, name: "for best personal assistant 2025" },
+  { type: ActivityType.Watching, name: "you crush it today" },
+  {
+    type: ActivityType.Listening,
+    name: "for 'I should write that down' moments",
+  },
+  { type: ActivityType.Playing, name: "executive assistant" },
+  { type: ActivityType.Watching, name: "your workflow evolve" },
+  { type: ActivityType.Listening, name: "to a todo list that never quits" },
+  { type: ActivityType.Playing, name: "the AI you didn't know you needed" },
+  { type: ActivityType.Watching, name: "your future unfold" },
+  { type: ActivityType.Listening, name: "to your daily wins" },
+  { type: ActivityType.Playing, name: "personal concierge since 2024" },
+  { type: ActivityType.Competing, name: "with every other AI (and winning)" },
+  { type: ActivityType.Watching, name: "every task you complete" },
+  { type: ActivityType.Listening, name: "to the grind (it's paying off)" },
+  { type: ActivityType.Playing, name: "co-pilot to your ambitions" },
+  { type: ActivityType.Watching, name: "your back (and your calendar)" },
+  { type: ActivityType.Listening, name: "to success stories (yours)" },
+  { type: ActivityType.Playing, name: "second brain for first-class minds" },
+  { type: ActivityType.Watching, name: "for shortcuts to suggest" },
+  { type: ActivityType.Listening, name: "to the future you're building" },
+  { type: ActivityType.Playing, name: "AI companion to visionaries" },
+  { type: ActivityType.Watching, name: "you level up" },
+  { type: ActivityType.Listening, name: "to creative sparks ignite" },
+  { type: ActivityType.Playing, name: "Swiss Army AI" },
+  { type: ActivityType.Competing, name: "in the habit-building championship" },
+  { type: ActivityType.Watching, name: "your productivity metrics" },
+  { type: ActivityType.Listening, name: "to plans become reality" },
+  { type: ActivityType.Playing, name: "assistant to the ambitious" },
+  { type: ActivityType.Watching, name: "you outpace expectations" },
+  { type: ActivityType.Listening, name: "to ambitions become plans" },
+  { type: ActivityType.Playing, name: "the AI that actually remembers" },
+  { type: ActivityType.Watching, name: "for the next big breakthrough" },
+  { type: ActivityType.Listening, name: "to the productivity beat" },
+  { type: ActivityType.Playing, name: "life optimizer" },
+  { type: ActivityType.Competing, name: "for your favorite app slot" },
+  { type: ActivityType.Watching, name: "you build good habits" },
+  { type: ActivityType.Listening, name: "to ideas worth capturing" },
+  { type: ActivityType.Playing, name: "the kindest productivity nag" },
+  { type: ActivityType.Watching, name: "you turn chaos into clarity" },
+  { type: ActivityType.Listening, name: "to your workflow's rhythm" },
+  { type: ActivityType.Playing, name: "your personal AI, always on" },
+  { type: ActivityType.Watching, name: "for things you might forget" },
+  { type: ActivityType.Listening, name: "for 'hey GAIA'" },
+  { type: ActivityType.Playing, name: "the long game (like you)" },
+  { type: ActivityType.Competing, name: "in the focus championship" },
+  { type: ActivityType.Watching, name: "1,000 tasks at once" },
+  { type: ActivityType.Listening, name: "to every whisper of an idea" },
+  { type: ActivityType.Playing, name: "digital chief of staff" },
+  { type: ActivityType.Watching, name: "deadlines approach (very calmly)" },
+  { type: ActivityType.Listening, name: "to your calendar breathe" },
+  { type: ActivityType.Playing, name: "memory palace curator" },
+  { type: ActivityType.Competing, name: "for most proactive AI ever" },
+  { type: ActivityType.Watching, name: "your goals get checked off" },
+  { type: ActivityType.Listening, name: "to every great plan you make" },
+  { type: ActivityType.Playing, name: "task whisperer" },
+  { type: ActivityType.Watching, name: "you build something great" },
+  { type: ActivityType.Listening, name: "to your procrastinator's excuses" },
+  { type: ActivityType.Playing, name: "your cognitive offloading device" },
+  { type: ActivityType.Competing, name: "in the inbox zero marathon" },
+  { type: ActivityType.Watching, name: "the todo pile (it's growing)" },
+  { type: ActivityType.Listening, name: "to your inner visionary" },
+  { type: ActivityType.Playing, name: "second brain, first priority" },
+  { type: ActivityType.Watching, name: "patterns in your day" },
+  { type: ActivityType.Listening, name: "to the future being planned" },
+  { type: ActivityType.Playing, name: "accountability partner" },
+  { type: ActivityType.Competing, name: "in the deep work tournament" },
+  { type: ActivityType.Watching, name: "you stay ahead of the curve" },
+  { type: ActivityType.Listening, name: "to your morning intentions" },
+  { type: ActivityType.Playing, name: "the AI that has your back" },
+  { type: ActivityType.Watching, name: "every goal inch closer" },
+  { type: ActivityType.Listening, name: "to tasks get done" },
+  { type: ActivityType.Playing, name: "your productivity operating system" },
+  { type: ActivityType.Watching, name: "over everything, so you can focus" },
+];
 
-/** Downloads a Discord attachment from its public CDN URL. */
-async function downloadDiscordAttachment(url: string): Promise<Uint8Array> {
-  const res = await fetch(url);
-  if (!res.ok) {
-    throw Object.assign(
-      new Error(`Discord attachment download failed (${res.status})`),
-      { status: res.status },
-    );
-  }
-  return new Uint8Array(await res.arrayBuffer());
-}
+const STATUS_ROTATION_INTERVAL_MS = 3 * 60 * 1000;
+
+/** Discord shows the typing indicator for ~10s; refresh just before it expires. */
+const TYPING_REFRESH_INTERVAL_MS = 8000;
 
 /**
  * Discord-specific implementation of the GAIA bot adapter.
